Updating README file and the second half of my original commit since I made a mistake writing the first commit
Added description to the README file. Linked the index page to all 3 recipe pages. Added an image to both the index and ghost brownies recipe page. Filled out the recipe page with ingredients and instructions using headers and lists. Left the other 2 recipe pages visitable but blank as I am low on time.